% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fake_data}
\alias{fake_data}
\title{Fake data for regression}
\format{data}
\usage{
fake_data
}
\description{
Fake data to show gaussian model example
}
\examples{
data('fake_data')
}
\keyword{datasets}
