% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_ConsRegArima.R
\name{plot.roll.ConsRegArima}
\alias{plot.roll.ConsRegArima}
\title{Plot an roll object
plot an roll.ConsRegArima object}
\usage{
\method{plot}{roll.ConsRegArima}(x, ...)
}
\arguments{
\item{x}{roll.ConsRegArima object}

\item{...}{Additional params passed to ggplot2::labs function}
}
\description{
Plot an roll object
plot an roll.ConsRegArima object
}
