% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ConsRegArima.R
\name{predict.ConsRegArima}
\alias{predict.ConsRegArima}
\alias{print.predict.ConsRegArima}
\alias{plot.predict.ConsRegArima}
\title{Predict function for ConsRegArima object}
\usage{
\method{predict}{ConsRegArima}(object, h = ifelse(is.null(newdata), 1,
  nrow(newdata)), newdata = NULL, intervals = 90, origdata = NULL)

\method{print}{predict.ConsRegArima}(x)

\method{plot}{predict.ConsRegArima}(x)
}
\arguments{
\item{object}{ConsRegArima object}

\item{h}{horizont to predict}

\item{newdata}{data frame in which to look for variables with which to predict.
In case there is no regression part, this parameter could be set NULL}

\item{intervals}{Confidence level for prediction intervals (default 90)}

\item{origdata}{Original data (default NULL). Useful if lagged predictive
variables are used in the formula}

\item{x}{object of class predict.ConsRegArima}
}
\value{
Returns an object of class predict.ConsRegArima

\item{predict}{dataframe with the predictions}
\item{table}{dataframe with the predictions as well as the fitted values}
\item{intervals}{Interval level}
\item{object}{original object}
}
\description{
Obtains predictions of ConsRegArima object
}
